requirements.txt
garminconnect==0.2.38
PyDrive2==1.16.5
fitparse==1.2.6
pandas==2.2.0
python-dateutil==2.8.2
python-dotenv==1.0.1
